found a little stupid bug guy sucker. Updated naming of array objects in response. Extracted fetching pongs to a seperate module. Added initial states to application a few application states
